:recycle: Added message repository
