+++ src/app.js
@@ -5,7 +5,7 @@ import handlebars from "express-handlebars";
 import { readFileSync } from "node:fs";
 import { Server as ServerIO } from "socket.io";
 import { config, connectDB } from "./config/config.js";
-import { messageManager } from "./dao/DBBasedManagers/ManagerSystem/ManagerSystem.js";
+import { messageService } from "./repositories/index.js";
 import cookieParser from "cookie-parser";
 import session from "express-session";
 import FileStore from "session-file-store";
@@ -97,9 +97,9 @@ io.on("connection", (socket) => {
   console.log("Client connected!");
 
   socket.on("addMessageEvent", async (data) => {
-    messageManager.addMessage(data);
+    messageService.addMessage(data);
 
-    const messages = await messageManager.getMessagesSortedByTimestamp();
+    const messages = await messageService.getMessagesSortedByTimestamp();
 
     io.emit("updateMessagesBoxEvent", messages);
   });
